requirements.txt
flake8
mypy
PyJWT
itsdangerous
black
pytest
isort


